requirements.txt
Flask==2.2.5
Authlib==1.2.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
itsdangerous==2.1.2